Add structured data to RESTFailure
In some cases, it can be helpful to have structured data about the
error that was returned from the REST API. Currently, all of the
information is added to the exception message, which makes it difficult
to access.

Add all of the information previously only included in the exception
message as structured data.

The exception message is kept the same to make this change backwards
compatible by only added new attributes to the exception and keeping
the message the same.